feat: move services to tutors-reader-lib
