+++ components/tutors-reader-lib/src/services/course-service.ts
@@ -1,5 +1,5 @@
 import path from "path-browserify";
-import { courseUrl, currentCourse, currentLo, currentUser, week } from "../../stores";
+import { courseUrl, currentCourse, currentLo, currentUser, week } from "tutors-course-reader/src/stores";
 import { replace } from "svelte-spa-router";
 import { Course } from "tutors-reader-lib/src/models/course";
 import { Lab } from "tutors-reader-lib/src/models/lab";
@@ -9,7 +9,7 @@ import { fetchUserById } from "tutors-reader-lib/src/utils/metrics-utils";
 import axios from "axios";
 import type { Lo, WeekType } from "tutors-reader-lib/src/types/lo-types";
 import type { Topic } from "tutors-reader-lib/src/models/topic";
-import { getKeys } from "../../environment";
+import { getKeys } from "tutors-course-reader/src/environment";
 
 export class CourseService {
   course: Course;
@@ -58,7 +58,7 @@ export class CourseService {
     const course = await this.getOrLoadCourse(courseId);
     currentCourse.set(course);
     courseUrl.set(course.url);
-    week.set(<WeekType>course?.currentWeek);
+    week.set(course?.currentWeek);
     await this.checkAuthenticated(course);
     await this.checkWhiteList(course);
     this.course = course;
