+++ sites/tutors-course-reader/src/navigators/NavBar.svelte
@@ -14,8 +14,8 @@
   import { getContext } from "svelte";
   import type { Course } from "tutors-reader-lib/src/models/course";
   import type { StudentMetric, User } from "tutors-reader-lib/src/types/metrics-types";
-  import type { MetricsService } from "../reader-lib/services/metrics-service";
-  import { PresenceService } from "../reader-lib/services/presence-service";
+  import type { MetricsService } from "tutors-reader-lib/src/services/metrics-service";
+  import { PresenceService } from "tutors-reader-lib/src/services/presence-service";
   import Icon from "tutors-reader-lib/src/iconography/Icon.svelte";
   import { AppBar, Divider } from "@brainandbones/skeleton";
