+++ sites/tutors-course-reader/src/navigators/sidebars/CalendarBar.svelte
@@ -2,7 +2,7 @@
   import { beforeUpdate, getContext } from "svelte";
   import type { Course } from "tutors-reader-lib/src/models/course";
   import type { Calendar, WeekType } from "tutors-reader-lib/src/types/lo-types";
-  import type { CourseService } from "../../reader-lib/services/course-service";
+  import type { CourseService } from "../tutors-reader-lib/src/services/course-service";
   import { calendarDrawer } from "../../stores";
 
   let course: Course = null;
