+++ sites/tutors-course-reader/src/pages/Note.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import { getContext, onMount } from "svelte";
   import { TopicNavigatorCard } from "tutors-ui";
-  import type { AnalyticsService } from "../reader-lib/services/analytics-service";
+  import type { AnalyticsService } from "tutors-reader-lib/src/services/analytics-service";
   import { revealSidebar } from "../stores";
   import * as animateScroll from "svelte-scrollto";
   import type { Lo } from "tutors-reader-lib/src/types/lo-types";
   import { NoteCard } from "tutors-ui";
-  import type { CourseService } from "../reader-lib/services/course-service";
+  import type { CourseService } from "tutors-reader-lib/src/services/course-service";
   import Loading from "./support/Loading.svelte";
   import Error from "./support/Error.svelte";
