+++ sites/tutors-course-reader/src/pages/Talk.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import { getContext, onMount } from "svelte";
   import { TalkCard, TopicNavigatorCard } from "tutors-ui";
-  import type { AnalyticsService } from "../reader-lib/services/analytics-service";
+  import type { AnalyticsService } from "tutors-reader-lib/src/services/analytics-service";
   import { revealSidebar } from "../stores";
   import * as animateScroll from "svelte-scrollto";
   import Loading from "./support/Loading.svelte";
   import Error from "./support/Error.svelte";
-  import type { CourseService } from "../reader-lib/services/course-service";
+  import type { CourseService } from "tutors-reader-lib/src/services/course-service";
   import type { Lo } from "tutors-reader-lib/src/types/lo-types";
 
   export let params: Record<string, string>;
