+++ sites/tutors-course-reader/src/pages/Video.svelte
@@ -2,13 +2,13 @@
   import { afterUpdate, getContext } from "svelte";
   import type { Lo } from "tutors-reader-lib/src/types/lo-types";
   import { TopicNavigatorCard, VideoCard } from "tutors-ui";
-  import type { AnalyticsService } from "../reader-lib/services/analytics-service";
+  import type { AnalyticsService } from "tutors-reader-lib/src/services/analytics-service";
   import { revealSidebar } from "../stores";
   import { querystring } from "svelte-spa-router";
   import * as animateScroll from "svelte-scrollto";
   import Loading from "./support/Loading.svelte";
   import Error from "./support/Error.svelte";
-  import type { CourseService } from "../reader-lib/services/course-service";
+  import type { CourseService } from "tutors-reader-lib/src/services/course-service";
 
   export let params: Record<string, string>;
   const analytics: AnalyticsService = getContext("analytics");
