+++ sites/tutors-course-reader/src/pages/Wall.svelte
@@ -3,11 +3,11 @@
   import type { Lo } from "tutors-reader-lib/src/types/lo-types";
   import { CardDeck, VideoCard } from "tutors-ui";
   import { currentCourse, currentLo } from "../stores";
-  import type { AnalyticsService } from "../reader-lib/services/analytics-service";
+  import type { AnalyticsService } from "tutors-reader-lib/src/services/analytics-service";
   import * as animateScroll from "svelte-scrollto";
   import Loading from "./support/Loading.svelte";
   import Error from "./support/Error.svelte";
-  import type { CourseService } from "../reader-lib/services/course-service";
+  import type { CourseService } from "tutors-reader-lib/src/services/course-service";
   import type Course from "./Course.svelte";
 
   export let params: Record<string, string>;
